+++ frontend/src/components/pure/ms-tab/index.vue
@@ -1,14 +1,15 @@
 <template>
   <a-tabs
     v-if="props.mode === 'origin'"
-    v-model:active-key="innerActiveKey"
+    v-model:active-key="tempActiveKey"
     :class="[props.class, props.noContent ? 'no-content' : '']"
+    @change="(val) => handleTabClick(val as string)"
   >
     <a-tab-pane v-for="item of props.contentTabList" :key="item.value" :title="item.label">
       <template v-if="props.showBadge" #title>
         <a-badge
           v-if="props.getTextFunc(item.value) !== ''"
-          :class="item.value === innerActiveKey ? 'active-badge' : ''"
+          :class="item.value === tempActiveKey ? 'active-badge' : ''"
           :max-count="99"
           :text="props.getTextFunc(item.value)"
         >
@@ -27,8 +28,8 @@
       v-for="item of props.contentTabList"
       :key="item.value"
       class="ms-tab--button-item"
-      :class="item.value === innerActiveKey ? 'ms-tab--button-item--active' : ''"
-      @click="innerActiveKey = item.value"
+      :class="item.value === tempActiveKey ? 'ms-tab--button-item--active' : ''"
+      @click="handleTabClick(item.value)"
     >
       {{ item.label }}
     </div>
@@ -39,12 +40,12 @@
   const props = withDefaults(
     defineProps<{
       mode?: 'origin' | 'button';
-      activeKey: string;
       contentTabList: { label: string; value: string }[];
       class?: string;
       getTextFunc?: (value: any) => string;
       noContent?: boolean;
       showBadge?: boolean;
+      changeInterceptor?: (newVal: string, oldVal: string, done: () => void) => void;
     }>(),
     {
       mode: 'origin',
@@ -54,9 +55,30 @@
     }
   );
 
-  const innerActiveKey = defineModel<string>('activeKey', {
+  // 实际值，用于最终确认修改的 tab 值
+  const activeKey = defineModel<string>('activeKey', {
     default: '',
   });
+  // 临时值，用于组件内部变更，但未影响到实际值
+  const tempActiveKey = ref(activeKey.value);
+
+  function handleTabClick(value: string) {
+    if (value === activeKey.value) {
+      return;
+    }
+    if (props.changeInterceptor) {
+      // 存在拦截器，则先将临时值重置为实际值（此时实际值是未变更之前的值），再执行拦截器
+      tempActiveKey.value = activeKey.value;
+      props.changeInterceptor(value, activeKey.value, () => {
+        // 拦截器成功回调=》修改实际值，也将临时值修改为实际值
+        activeKey.value = value;
+        tempActiveKey.value = value;
+      });
+    } else {
+      // 不存在拦截器，直接修改实际值
+      activeKey.value = value;
+    }
+  }
 </script>
 
 <style lang="less" scoped>
